// frontend/src/api/generated/api.schemas.ts
export interface PoLineItemInput {
  description: string;
  quantity: string;
  /** @nullable */
  quotation_line_item_id?: string | null;
  unit: string;
  unit_rate: string;
}